DONE fix AuthProvider
