+++ frontend/src/App.js
@@ -5,6 +5,14 @@ import { createStackNavigator } from "@react-navigation/stack";
 
 import Title from "src/containers/Title";
 
+import firebase from "firebase";
+import { AuthProvider } from "src/utils/auth";
+import { config } from "src/utils/config";
+
+firebase.initializeApp(config);
+const db = firebase.firestore();
+const auth = firebase.auth();
+
 const Stack = createStackNavigator();
 
 const StackNavigatorProps = {
@@ -16,9 +24,11 @@ const StackNavigatorProps = {
 const App = () => (
   <>
     <NavigationContainer>
+      <AuthProvider auth={auth}>
       <Stack.Navigator {...StackNavigatorProps}>
         <Stack.Screen name="Title" component={Title} />
-      </Stack.Navigator>
+        </Stack.Navigator>
+        </AuthProvider>
     </NavigationContainer>
   </>
 );
